requirements.txt
rembg
Pillow
flask
gunicorn
